Add location reducer
This is just a placeholder, to make sure that we have a default
location on our state. This should always refer to the current
location, and should probably have metadata like HDOP. We should have a
location module fire a “UPDATE_LOCATION” action when location updates,
but for now this will do.