Geolocation in progress
Why is this in progress and not done? Because of 3 things:

1 - The new location store is breaking all the code that depends on an
ID. That will take some (albeit small)  thinking to fix, because we abolished a default
location

2 - Testing in Firefox, there is a prompt for checking
whether I want to grant geolocation permissions for the website. Okay,
fine. But that means the component goes through the ComponentDidMount
and does not do anything. Which is, well, weird, and I want to test a
little bit more.

3 - I still have to do the dispatching the action part. No big deal,
just match the payload to the stuff in the reducers.