Fixes invariants and an if condition
The if condition was on an property of the error variable, but the error
variable is a bool. So I changed the conditional to the payload.

Also, the invariants where being called when redux mounts the reducers,
and failing. So the invariants are now inside the switch statement.

Last but not least, invariants are now OR instead of AND