feature/observation-api fixes (#42)
* fix rnnodeapp build script

By ignoring an Electron-specific package, "original-fs", leftover from
mapeo-mobile-server

* retry localhost requests

While rnnodeapp is still starting up, the frontend may be too eager and
send a request to it. This is a race condition, and the frontend's
request will fail. This commit adds a quick and weak fix to it, by
retrying failed requests at least 3 times. Often this is enough to
create some waiting time until rnnodeapp is up.