Flatten app state
Previously the state was all under a prop, app.
This made it hard to whitelist or blacklist props for redux-persist.
Redux-persist was overwriting the GPS status, because the GPS is called outside the persist gate.
With larger preset files with complex svgs, persisting the entire state synchronously was causing perf issues