adds inset map view
running into strange bug though with the MapboxGL SymbolLayer component (that should load a custom icon). It crashes the app. The only error in the logs is `abort_message: assertion "terminating with uncaught exception of type jni::PendingJavaException" failed`... If you swap this component with the CircleLayer component, it runs fine. @gmaclennan can you see if you're getting this error as well?