feat: Upgrade to React Native v0.60.4 (#183)
Updating to latest React Native (from our current v0.58) gives us lots of benefits:

Fixes to bugs with JSC (JavaScriptCore) which were causing crashes with Samsung S7 phones and other phones (we had some hacky workarounds)
Ready for 64-bit versions which will be required for listing on Google Play after August 2019.
Support for new React features (react hooks which make the code easier to debug and more concise
It's now much easier to use native modules - this means things like the camera and gps - without complicated integration. This should mean that building for iOS should be much easier
It uses AndroidX which is a redesigned module system for Android with clearer naming and versions, which should make compatibility and upgrading a lot less bug-prone.
More info in the announcements of RN v0.59 and RN v0.60

The latest version of nodejs-mobile (which we use for Mapeo Core) has not been working for Mapeo, and we needed to do some hacky fixes for bugs we were hitting which were already fixed in the latest version. Now we're running latest nodejs-mobile, so we get the bug fixes from there.