Fix: Fix crash from switching away from app during camera view
We've seen several errors in bugsnag caused by subscriptions and the expo-sensors module.
Turns out we were doing too much, trying to unsubscribe from the accelerometer when the app was going into the background. Expo was already doing that, so we were throwing an error by doing it twice. Re-writing this as a function component helped clear up the logic and identify the bugs.
fixes #330 #187 #138