fix(ObservationView): Fix display of observation icon on inset map
Switch to render a react native Image as the icon instead of rendering
a symbol layer in the map, which was causing some rendering issues on
certain devices. Since the map is static (cannot be panned or zoomed) we
can just render the image in the center of the map and it will be in the
correct location.