feat: Do not ship Mapeo for ICCAs assets with Mapeo (#447)
* feat: Create ICCAs build scripts & exclude ICCA screens from main build

Use Metro build config to conditionally include ICCA screens so that
ICCA resouces are not included in the main Mapeo build. This avoids the
images used in the ICCAs intro screens increasing the file size of the
main Mapeo APK.

* feat: Ship apk variants with only the presets they need

Choose the presets that are included in the apk at build time instead of
runtime. This reduces the apk size and removes runtime code for
conditionally running code based on app variant.

* chore: Patch react-native so gradle tasks package ICCA variant

Adds the app variant name as an environment variable in the gradle build
script so that metro bundler can package JS files according to variant
by using the `sourceExts` option.