Update backend build script to only copy whitelisted files
The backend build script uses noderify to bundle JS into a single file.
Previously the script would then blacklist unnecessary files to delete
them from the packaged app, but that is a maintenance burden.
This change whitelists the necessary files that must be included in the
APK and only includes those.