Update settings store key for new settings shape
This will cause anyone who has used the "Coordinate format" setting to
loose their chosen format, but we have not released that yet. For the
next time we update the settings that are persisted, we need a way to
migrate old settings, because leaving this at @1 broke my phone, because
it had stored settings that did not have the experiments option.