fix: upload release apk with armeabi-v7a and arm64-v8a ABIs
Previously we were uploading the `.aab` with all ABIs to Google Play.
The Play Store generates a custom apk from this `.aab` with only the
ABI that the device supports. This causes p2p update to fail when
updating between devices with different ABIs. By uploading a `.apk`
instead of a `.aab`, Play Store will not generate custom apks. The
downside of this is a larger install from Google Play (the apk
containing both the armeabi-v7a and arm64-v8a ABIs), but it avoids
issues with p2p updates. Note: This means that Intel devices cannot
install from the Google Play Store (since we are no longer shipping x86
and x86_64 ABIs). This will be resolved in a future commit, and only
affects a very small number of phones (it also affects Chromebooks that
support Android apps and run on x86).