chore: don't run install scripts on backend build
This is just a DX issue to save time and unnecessary downloads during the
backend build. The install / postinstall scripts all seem to be for
downloading or compiling native builds, but done during install this will be
the native builds for the dev machine arch. We will be rebuilding these anyway
for mobile, so no need to run these scripts. Ensure we start in the right place